backend/requirements.txt
daphne==4.1.2
Django==4.2
django-jazzmin==2.6.0
django-tenants==3.6.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.2
django-environ==0.11.2
psycopg2==2.9.9
whitenoise==6.1.0
django-extensions==3.2.3
